Loan details page UI updated
